TMS Simulation with figures
The goal is to validate the TMS. This is done by giving a test signal, in this case it is a step function that steps to a high temp value then drops off. The figures generated from the data collected in the MIL and SIL models that the data produced match and produce the desired outputs.
POG